added new function for detecting collision
the checkCollision function checks to see if the ball has collided
with one of the four walls of the canvas. If it has it updates
xSpeed or ySpeed as appropriate to make the ball bounce of the wall.

We need to know where the left, right, top, and bottom edges are - to
determine if these edges have exceeded the bounds of the the playing area.
We group the values in an object called - ball - that has left, right, top,
and bottom properties.
Identifying the left and top ball edges is easy: they're ballPosition.x and
ballPosition.y, respectively.
To get the right and bottom edges, we add BALL_SIZE to ballPosition.x and
ballPosition.y