added two objects; leftPaddle and rightPaddle
the two objects contain the edges of their respective paddles
as four properties, left, right, top, and bottom. Determining
where the edges of the right paddle are requires a bit more math
because we have to take into account the width of the canvas, the
offset of the paddle, and the width of the paddle.