defined an function to check collisions for paddle and ball
the functions takes the ball object and one of the paddle objects
and returns true if the ball is intersecting with that paddle. This
function uses a long Boolean expression made up of four subexpressions
that are all &&'d together, so it returns - true - only if all four
subexpressions are true.