created a adjustAngle function and updated the checkCollision function
the adjustAngle function has two parameters, -distanceFromTop- and -distanceFromBottom-
These represent the distance from teh top of the ball to top of the paddle and from the
bottom of the paddle to the bottom of the ball, respectively.

Next we updated the checkCollision function to call the new adjustAngle function as part
of the collision detection logic for the two paddles.