defined a new function -> followBall()
the new function applies some rudimentary artifical intelligence
to move the computer's paddle. Added between the draw() and update()

Within the followBall() function, we define objects representing the ball
and the left paddle, each with -top , bottom- properties representing their
upper and lower bounds.
Then we implement the paddle movement logic with two if statements - if the
top of the ball is above the top of the paddle, we move the paddle up by subtracting
MAX_COPUTER_SPEED from leftPaddleTop - Likewise, if the bottom of the ball is below
the bottom of the paddle, we move the paddle down by adding MAX_COMPUTER_SPEED
to leftPaddleTop

We call our new followBall() function within the -update- function. This way,
moving the left paddle becomes part of the process of updating the state of the game
that happens with each iteration of the game loop.