declared our first class called GameView
This class is responsible for the player's view of the game, or how the game is displayed.
Because the game uses a canvas for rendering, the GameView is also responsible for drawing
things to the canvas, such as the ball and paddles, and for displaying the -Game Over- text.

The -GameView- constructor gets a reference to the canvas and its drawing context and saves
these as properties called canvas and ctx. It also stores some values that will be used for
drawing: the width and height of the canvas, and the offset of the canvas from the top of the
browser viewport.

The -draw- method uses rest parameters, this way, you can pass multiple arugments to -draw-, and
all the arguments will be collected into a single array called entities. - Each argument will be an
object representing one of the elements in the game: the ball and the two paddles.
The method first draws a black rectangle to clear the canvas and then goes through the array of
elements, calling each element's own -draw- method in turn - and passing the drawing context as
an argument. This will work only if every object passed to -GameView.draw- has its own draw method.

The -draw- method on -GameView- has the ultimate responsibility for drawing to the canvas with each
repetition of the game loop, -- but it delegates responsibility for actually drawing the game elements
to the objects representing those elements. -- In effect, each element in the game -knows- how to draw
itself, and -GameView.draw- just orchestrates the calls.

The -drawScores- method takes an object containing the two scores and draws them to the canvas. Instead
of relying on a global variable for the width of the canvas, it's able to use the width property from
the -GameView- class by referring to -this.width-

The -drawGameOver- method is getting the width and height from -GameView- rather than from global variables.