created a superclass called -Entity-
This superclass will be a parent to the subclasses Paddle and Ball.

The Entity class exists to share the general code common to both the paddles
and the ball. This includes code for keeping track of the sizes and positions
of the elements, calculating the boundaries of the elements for collision detection,
and drawing the elements. - Since all the game elements are rectangles, all this
code is the same whether we're dealing with a paddle or the ball.

This shows the beauty of Object-Oriented Programming: we can write all the common
code once in the superclass, and let the subclasses inherit it.

The -Entity- constructor takes an x- and a y-coordinate for the top-left corner of
the entity, and a width and height for its size. These are saved as properties.

The -boundingBox- method returns an object with teh left, right, top, and bottom
bounds of the entity.

The -draw- method takes a drawing context and draws a white rectangle using the
properties defined in the constructor. -- The objects passed  into the draw method
on -GameView- will all be subclasses of Entity, and it's the -draw- method on -Entity-
that will be called for each item.