declared class called -Paddle- that extends -Entity- class
this class includes three -static properties- which are properties assigned
to the class itself rather than to an individual instance of the class.

A static property's value will be shared across all instances of the class.
-- In this case, while each instance of Paddle needs its own x- and y-coordinates,
every -Paddle- object should have the same width and height, as well as the same
offset from the left or right edge of the canvas.

Static properties are accessed using dot notation, as with an instance's properties,
except that you us the class name on the left side of the dot, rather than -this- or
the name o fthe instance. For example, -Paddle.WIDTH- accesses the WIDTH static property.

The -Paddle- constructor has only two parameters: x and y. It uses - super - to call the
-> constructor of its superclass (Entity) <- passing through the x and y parameters as
well as Paddle.WIDTH for the width parameter and Paddle.HEIGHT for the height parameter.