created the -Ball- class that extends the -Entity class
This class has a static property called SIZE that defines the width and height
of the ball. -- Next comes its constructor method, the first thing the -Ball-
constructor does is call the constructor of its superclass, -Entity-, this time
passing 0 for the x and y parameters and Ball.SIZE for the width and height
parameters. The 0s are just placeholders; -- in fact, the ball starts each point
at the same position, (20,30). This position is handled by the -Ball- class's
-- -init- method, which is called for the first time from the constructor.

The -init- method itself is defined to set the initial position and speed of
the ball. This method will be called whenever the ball needs to be reset to
its initial position (after a point is scored).

The next method, -update- uses the ball's current speed to update its
x- and y-position.

The -adjustAngle- method changes the ball's vertical speed depending on where
the ball hits the paddle.