declared two methods to check for collision
The -checkPaddleCollision- method takes two parameters: an object representing one of
the paddles, and xSpeedAfterBounce. The latter represents the new value we should set
xSpeed to if a bounce off one of the paddles occurs, and it lets us configure whether
the ball is expected to always bounce to the right (from the left paddle) or to the
left (from the right paddle). -- We enforce that a collision with the left paddle makes
the ball bounce right and vice versa to avoid weirdness where the ball could bounce
around - inside - the paddle.

We use -boundingBox- method from the parent -Entity- class to get the bounding boxes
of the ball and the paddle, storing them as -ballBox- and -paddleBox-

Next, we compare the various bounding box edges to determine if a collision has taken
place between the ball and the paddle, saving the result in the Boolean variable
-collisionOccured- -- If collisionOccurred is TRUE, we call the -adjustAngle- method
with the appropriate distances as determined from the bounding boxes, then set the
ball's xSpeed and xSpeedAfterBounce.

Finally, the -checkWallCollision- method checks to see if a collision has occurred
between the ball and a wall. It takes the width and height of the playing area
and an object representing the scores as parameters. -- If the ball hits the left wall
or right wall, the appropriate score is incremented, and the ball is reset with a
call to the -init- method. If it hits the top or bottom wall, it bounces.