declared a -Scores- and -Computer- class
The -Scores- class is a simple container for keeping track of the current scores.
The -Computer- class contains the logic for following the ball.

The -Scores- constructor sets the scores for the left and right players to 0.

The -Computer- class has a single method called -followBall- which updates
the left paddle's position based on the ball's position. It's a - static method -
meaning it doesn't need an instance of the class to be called. -- Static methods
are called using the class name rather than the instance name, like this:
 - Computer.followBall(leftPaddle, ball) -
NOTE - the -Computer- class doesn't have any properties, so we don't need to
create an instance of it. Since the -Computer- class is never instantiated, it
also doesn't need a constructor.