created the class object -> Game
the -Game- class is where all other classes get instantiated (where applicable)
and are stitched together and orchestrated.

The -Game- constructor first instantiates the GameView, Ball, and Paddle classes.

The leftPaddle instance is created with -Paddle.OFFSET- for its x-coordinate. The right one
uses -Paddle.OFFSET, Paddle.WIDTH, and this.gameView.width- to determine its x-coordinate, similar
to the previous chapter.

Instantiating other classes inside a class is a common feature of Object-oriented code. This
technique is called -> composition <- becuase we're composing instances inside other instances.

Next, the -Game- constructor instantiates -Scores- and sets the -gameOver- Boolean to false. Finally
it sets up a -mousemove- event listener to update the right paddle's position when the user moves the mouse.
An event listener set up in a class constructor works just like the other event listeners we've seen: it will
be available as long as the application is running, and it triggers its handler function whenever the event
is detected.

After the constructor we have the -Game- class's -draw- method, which has the top-level responsibility for
drawing all the visual aspects of the game.
First the method calls -this.gameView.draw- passing the three main game elements, -this.ball-, -this.leftPaddle-,
and -this.rightPaddle-. This is a call to the draw method of the -GameView- class which took a variable number of
objects as arguments and called the draw method on each one.
The net result is that -game.draw- calls -gameView.draw-, which calls -ball.draw-, -left.Paddle.draw-, and -right.Paddle.draw-.
It may seem a bit roundabout, but you'll often find cases like this in Object-oriented code, where keeping the code in its
logical place requires jumping through some hoops.
In this case, -game.draw- is responsible for knowing WHICH objects get drawn (because the Game class keeps track of all the
game elements); -gameView.draw- is responsible for the drawing context, clearing the canvas, and calling the -draw- methods
on the elemnts; and the draw method on each game element is responsible for knowing how to draw itself.

After it has drawn all the entities, the draw method calls -this.gameView.drawScores- passsing the -this.scores- object.