continued creating more methods in -Game- class
The -Game- class's checkCollision method coordinates all the collision detection logic. First it calls
the ball's -checkPaddleCollision- method twice, to check for collisions between the ball and each of the
paddles. (Recall that this method takes two arguments: a -Paddle- object and a new post-bounce value for
xSpeed.) For the left paddle, we know that we want the ball to bounce right, so we make the new -xSpeed-
positive by taking the -Math.abs- of the current -xSpeed-.
For the right paddle we want the ball to bounce left, so we make the new -xSpeed- negative by negating
the result of -Math.abs(xSpeed)-

Next, the -checkCollision- method calls -ball.checkWallColision- to handle wall collisions. This method
takes the width and height (because the -Ball- object doesn't know how big the playing area is) and scores
(so they can be incremented if a side is hit). Finally
the method checks to see if their score has exceeded the threshold and, if so, sets -this.gameOver- to TRUE.

The -Game- object's -update- method controls the changes of state between each repetition of the game loop.
It calls the ball's update method to move the ball, then tells the computer to move the left paddle based on
the ball's new position using the -Computer.followBall- static method.

The last method of the -Game- class, loop, defines the game loop. We call -this.draw-, -this.update-, and
-this.checkCollision- in sequence. Then we check to see if -this.gameOver- is TRUE. If so, we call -draw-
again to render the final score, and we call -gameView.drawGameOver- to render the [GAME OVER] text. Otherwise
we use -setTimeOut- to call the -loop- method again after 30 ms, to continuing the game.